feat: Add featured products and dynamic product rendering on home page
This commit introduces the display of featured products on the home page. This feature highlights select products, providing more visibility and promoting user engagement.

In addition, dynamic rendering of other products has been implemented on the home page. This change allows for a more flexible and scalable product display, as products can be added or removed from our database and the changes will be reflected on the home page in real time.